# Hack around world age issue PyPlot (fix #734)