@mkborregaard 2017-05-19
Hi @darwindarak thanks for taking the initiative on this. It is possible that we should start by merging this if it does indeed temporarily resolve the issue, although it'd be nice to get to the bottom of it instead. Do you have any insights you can share so we could work together on resolving this?