@darwindarak 2017-05-20
@mkborregaard I think that's a good idea.  We can open up another PR to improve on this fix.

There should be a way to fix this without using `invokelatest`.  If we only merge the changes to `backends/pyplots.jl` (and not `pipeline.jl`), then we can call
```julia
using Plots
backend() # need this
plot(rand(10))
```
without any errors or we can call 
```julia
using Plots
plot(rand(10)) # encounters the world age problem
plot(rand(10)) # works fine after this
```
and only get the error on the first call to `plot`.  I haven't had the time to dig any further than that.