@mkborregaard 2017-05-20
I guess this means it would work for most people if we changed the default backend to be something different from pyplot, necessitating a `pyplot` call. Would you mind splitting this PR into two? We could probably merge the one into pyplot immediately.