@darwindarak 2017-05-20
Ok, I've split the first part into #848 and rebased the current PR on top of that one.