@mkborregaard 2017-06-08
@darwindarak it looks like this commit wasn't properly rebased, it still contains the bit from #848 - could you rebase again?