@mkborregaard 2017-06-08
I'd be inclined to merge this and get on with it. Comments, @daschw  @jheinen @pkofod @tbreloff ?