@mkborregaard 2017-06-08
Unfortunately, it doesn't appear to fix this issue: https://github.com/JuliaPlots/Plots.jl/issues/745#issuecomment-294122675
But that issue is fixed by calling the backend explicitly in the script, e.g. 
```
using Plots
gr()
display(plot(rand(4,4)))
```
I think we might search that way for a solution instead.