@randy3k 2017-06-12
I need this when running `pyplot` on jupyterhub.

```
using Plots
ENV["MPLBACKEND"] = "Agg"
pyplot()
```
The second line ensures that no gui is opened.