@mkborregaard 2017-08-21
I just reran the benchmarks above comparing to current master, and found only a very slight performance reduction. So, unless anyone opposes this (@tbreloff @daschw @ChrisRackauckas @pkofod ) I'd like to just merge this, even if the problem only exists in Juno.

@darwindarak Would you rebase and remove the WIP in the title?