@mkborregaard 2017-08-21
@randy3k can you see if this helps: https://github.com/JuliaPlots/Plots.jl/pull/1030 ?