@darwindarak 2017-08-24
Rebased to master