@mkborregaard 2017-08-24
Thanks @darwindarak 
For reference, these are my timings:
```julia
# on current master
julia> using Plots, BenchmarkTools
julia> backend()
Plots.GRBackend()

julia> plot(rand(10))
julia> @benchmark plot($rand(10))
BenchmarkTools.Trial:
  memory estimate:  129.53 KiB
  allocs estimate:  2294
  --------------
  minimum time:     487.161 μs (0.00% GC)
  median time:      605.256 μs (0.00% GC)
  mean time:        648.155 μs (4.43% GC)
  maximum time:     7.473 ms (84.87% GC)
  --------------
  samples:          7643
  evals/sample:     1

julia> pyplot()
Plots.PyPlotBackend()

julia> plot(rand(10))
julia> @benchmark plot($rand(10))
BenchmarkTools.Trial:
  memory estimate:  129.88 KiB
  allocs estimate:  2305
  --------------
  minimum time:     763.622 μs (0.00% GC)
  median time:      790.258 μs (0.00% GC)
  mean time:        862.438 μs (3.53% GC)
  maximum time:     7.667 ms (76.62% GC)
  --------------
  samples:          5765
  evals/sample:     1



# With this PR:
julia> backend()
Plots.GRBackend()

julia> plot(rand(10))

julia> @benchmark plot($rand(10))
BenchmarkTools.Trial:
  memory estimate:  129.48 KiB
  allocs estimate:  2293
  --------------
  minimum time:     485.927 μs (0.00% GC)
  median time:      507.858 μs (0.00% GC)
  mean time:        576.974 μs (4.42% GC)
  maximum time:     7.094 ms (84.36% GC)
  --------------
  samples:          8585
  evals/sample:     1

julia> pyplot()
Plots.PyPlotBackend()

julia> plot(rand(10))
julia> @benchmark plot($rand(10))
BenchmarkTools.Trial:
  memory estimate:  129.83 KiB
  allocs estimate:  2304
  --------------
  minimum time:     778.650 μs (0.00% GC)
  median time:      850.840 μs (0.00% GC)
  mean time:        995.671 μs (3.36% GC)
  maximum time:     10.333 ms (83.98% GC)
  --------------
  samples:          4985
  evals/sample:     1
```
The possible regression is neglible IMHO.